refactor: rename package's name
