+++ scripts/evaluate.py
@@ -12,8 +12,8 @@ from torchmetrics.functional import confusion_matrix, accuracy, auroc, f1_score
 from torchvision.transforms import transforms as T
 from torchvision.transforms._transforms_video import ToTensorVideo
 
-from TubeViT.dataset import MyUCF101
-from TubeViT.model import TubeViTLightningModule
+from tubevit.dataset import MyUCF101
+from tubevit.model import TubeViTLightningModule
 
 
 @click.command()
