+++ scripts/train.py
@@ -9,8 +9,8 @@ from torch.utils.data import DataLoader, RandomSampler
 from torchvision.transforms import transforms as T
 from torchvision.transforms._transforms_video import ToTensorVideo
 
-from TubeViT.dataset import MyUCF101
-from TubeViT.model import TubeViTLightningModule
+from tubevit.dataset import MyUCF101
+from tubevit.model import TubeViTLightningModule
 
 
 @click.command()
