+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import find_packages, setup
 
 # yapf: disable
 setup(
-    name='TubeViT',
+    name='tubevit',
     packages=find_packages(),
     version='0.1.0',
     description='An unofficial implementation of TubeViT in "Rethinking Video ViTs: Sparse Video Tubes for Joint Image and Video Learning".',
